Manual CSS and HTML updated
The manual.html file is going to have more text changes shortly. with a new commit.  This commit is being made now, since it's file changes can be used outside of this branch.  Future modifications to manual.html in this branch will be more drastically changed with updated graphics to coincide with this branch's new UI.

Changes in this commit include:
* CSS was modified to using transition animations for aside
* nav TOC was removed since it was duplicitous to aside.  Also, removing nav TOC makes landing area more aesthetically pleasing.
* Text was updated in some sections